These changes should reflect the correct signs.
Both the Stokes drift and its derivative wrt depth are computed as positive numbers. When they enter the kernel within "onshore_transport_kernel" they get a minus sign. We know that velocities get a minus sign in convection-diffusion equations, so this should be correct. The advection, diffusion and production terms that appear in both ActiveLayer.jl as well as in ActiveLayerOnshore.jl have the same signs.