Pruning of "AdvectionSpec.jl"
This is the result of pruning "AdvectionSpec.jl" - which was mostly based on code from the "onshore-run.jl" notebook in the "115-initial-sediment-parameter" branch - under the guidance of @jhidding on June 4.
The idea is that only the essentials for unit testing remain and that the new "transport_test_input" function is used.